Optimize handling of very large slack teams. Fixes #695
Stop getting users if we reach 2000 users. Slack will rate-limit us
even if we follow their limits.
This means that we now have to lookup every user that says a message
for the first time. This should be less intensive on the API.

This also disables partly f31ba861996fa81ee3e9449549d4b92bb1b3778d for now.
ChannelMembers will not be filled.